refactor: deploy mojo's to own namespace
