+++ src/main/java/ch/ivyteam/ivy/maven/deploy/AbstractDeployMojo.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ch.ivyteam.ivy.maven;
+package ch.ivyteam.ivy.maven.deploy;
 
 import java.io.File;
 import java.io.IOException;
@@ -29,7 +29,9 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 import org.apache.maven.shared.filtering.MavenFileFilter;
 
-import ch.ivyteam.ivy.maven.DeployToEngineMojo.DefaultDeployOptions;
+import ch.ivyteam.ivy.maven.AbstractIntegrationTestMojo;
+import ch.ivyteam.ivy.maven.IarPackagingMojo;
+import ch.ivyteam.ivy.maven.deploy.DeployToEngineMojo.DefaultDeployOptions;
 import ch.ivyteam.ivy.maven.engine.deploy.DeploymentOptionsFileFactory;
 import ch.ivyteam.ivy.maven.engine.deploy.YamlOptionsFactory;
 import ch.ivyteam.ivy.maven.engine.deploy.dir.FileDeployer;
