+++ src/main/java/ch/ivyteam/ivy/maven/deploy/DeployToEngineMojo.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ch.ivyteam.ivy.maven;
+package ch.ivyteam.ivy.maven.deploy;
 
 import java.io.File;
 import java.nio.file.Paths;
